requirements.txt
fastapi
uvicorn
tortoise-orm[asyncpg]
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-multipart